client-day-17

+++ controllers/users.js
@@ -27,12 +27,20 @@ router.post("/add", async (req, res) => {
       email: req.body.email.toLowerCase(),
       phoneNumber: req.body.phoneNumber,
       password: await bcrypt.hash(req.body.password, 10),
-      type: req.body.type,
       createdOn: new Date()
     }
 
-    if(req.body.type === userTypes.USER_TYPE_STANDARD)
-      record.departmentId = req.body.departmentId;
+    if(req.user.type === userTypes.USER_TYPE_STANDARD)
+    {
+      record.departmentId = req.user.departmentId;
+      record.type = userTypes.USER_TYPE_STANDARD;
+    }else
+    {
+      record.type = req.body.type;
+      if(req.body.type === userTypes.USER_TYPE_STANDARD)
+        record.departmentId = req.body.departmentId;
+    }
+  
 
     const user = new User(record)
 
@@ -54,6 +62,8 @@ router.post("/edit", async (req, res) => {
     const user = await User.findById(req.body.id);
     if (!user) throw new Error("User does not exists");
 
+    if(req.user.type === userTypes.USER_TYPE_STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+      throw new Error('invalid request');
 
     const record = {
       name: req.body.name,
@@ -77,7 +87,7 @@ router.post("/edit", async (req, res) => {
 });
 
 
-router.delete("/delete", async (req, res) => {
+router.post("/delete", async (req, res) => {
   try {
     if (!req.body.id) throw new Error("User id is required");
     if (!mongoose.isValidObjectId(req.body.id))
@@ -85,6 +95,12 @@ router.delete("/delete", async (req, res) => {
 
     const user = await User.findById(req.body.id);
     if (!user) throw new Error("User does not exists");
+
+    if(req.body.id === req.user._id.toString())
+        throw new Error('invalid request');
+
+    if(req.user.type === userTypes.USER_TYPE_STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+      throw new Error('invalid request');
 
     await User.findByIdAndDelete(req.body.id);
